requirements.txt
streamlit
torch
torchvision
ultralytics
opencv-python-headless
Pillow
numpy
dill
folium 
streamlit-folium
branca
